feat: support passing MongoDB client connection options
This library is printing all sorts of mongoose warnings to the console. We can't fix it
because the library currently provides no way to thread mongoose/mongo connection options.

This resolves that.

We decided to make this a breaking change, since there's no particular reason to support
the current share, the breaking change is very easy to fix, and at least at Good Eggs
we'll only have a couple places to update.